requirements.txt
dagger==1.3.0
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.26.0
